python 3.10: explicitly cast some numbers to int
These currently raise deprecation warnings, but will no longer get
autoconverted in python 3.10

- Some Qt functions expect int, not float, and got truncated automatically.
- The QSocketNotifier is sip.voidptr wrapping an int file descriptor

+++ src/calibre/gui2/store/search/results_view.py
@@ -31,7 +31,7 @@ class ImageDelegate(QStyledItemDelegate):
                 dpr = img.devicePixelRatio()
                 scaled, nw, nh = fit_image(img.width(), img.height(), w, h)
                 if scaled:
-                    img = img.scaled(nw*dpr, nh*dpr, Qt.AspectRatioMode.IgnoreAspectRatio, Qt.TransformationMode.SmoothTransformation)
+                    img = img.scaled(int(nw*dpr), int(nh*dpr), Qt.AspectRatioMode.IgnoreAspectRatio, Qt.TransformationMode.SmoothTransformation)
             iw, ih = int(img.width()/dpr), int(img.height()/dpr)
             dx, dy = (option.rect.width() - iw) // 2, (option.rect.height() - ih) // 2
             painter.drawPixmap(option.rect.adjusted(dx, dy, -dx, -dy), img)
